requirements.txt
asgiref==3.2.7
certifi==2019.11.28
chardet==3.0.4
decorator==4.4.2
Django==3.0.4
djangorestframework==3.11.0
idna==2.9
imageio==2.8.0
imageio-ffmpeg==0.4.1
moviepy==1.0.0
mysqlclient==1.4.6
numpy==1.18.2
Pillow==7.0.0
pkg-resources==0.0.0
proglog==0.1.9
pytz==2019.3
requests==2.23.0
sqlparse==0.3.1
tqdm==4.43.0
urllib3==1.25.8
